+++ chrono_hil/ROM/driver/ChROM_PathFollowerDriver.h
@@ -22,13 +22,14 @@ class ChROM_PathFollowerDriver {
 public:
   ChROM_PathFollowerDriver(std::shared_ptr<Ch_8DOF_vehicle> m_rom,
                            std::shared_ptr<ChBezierCurve> curve,
-                           float target_speed, double look_ahead_dist,
-                           float PID_st_kp, float PID_st_ki, float PID_st_kd,
-                           float PID_sp_kp, float PID_sp_ki, float PID_sp_kd);
+                           double target_speed, double look_ahead_dist,
+                           double PID_st_kp, double PID_st_ki, double PID_st_kd,
+                           double PID_sp_kp, double PID_sp_ki,
+                           double PID_sp_kd);
 
-  void Advance(float step);
+  void Advance(double step);
   DriverInputs GetDriverInput();
-  void SetCruiseSpeed(float target_speed);
+  void SetCruiseSpeed(double target_speed);
 
 private:
   std::shared_ptr<ChBezierCurve> m_curve;
@@ -36,27 +37,27 @@ private:
   std::shared_ptr<Ch_8DOF_vehicle> m_rom;
 
   // target spped
-  float m_target_speed;
+  double m_target_speed;
 
   // steering PID control
-  float m_st_kp;
-  float m_st_ki;
-  float m_st_kd;
+  double m_st_kp;
+  double m_st_ki;
+  double m_st_kd;
 
   // speed PID control
-  float m_sp_kp;
-  float m_sp_ki;
-  float m_sp_kd;
+  double m_sp_kp;
+  double m_sp_ki;
+  double m_sp_kd;
 
   // steering errs
-  float m_st_err;   // cached error
-  float m_st_err_d; // d_err
-  float m_st_err_i; // i_err
+  double m_st_err;   // cached error
+  double m_st_err_d; // d_err
+  double m_st_err_i; // i_err
 
   // speeed errs
-  float m_sp_err;   // cached error
-  float m_sp_err_d; // d_err
-  float m_sp_err_i; // i_err
+  double m_sp_err;   // cached error
+  double m_sp_err_d; // d_err
+  double m_sp_err_i; // i_err
 
   // look ahead distance
   double m_dist;
